requirements.txt
openai-whisper
torch
numpy
customtkinter
packaging